requirements.txt
asgiref==3.8.1
Django==5.1.4
django-environ==0.11.2
dj-database-url==0.5.0
gunicorn>=20.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
whitenoise==6.8.2